Refactor: move the data density graph to its own file
